pyproject.toml
[tool.poetry]
name = "cedazo"
version = "0.1.0"
description = "Transformation of Excell files"
authors = ["Iwan van der Kleijn <iwanvanderkleijn@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
takeme-openpyxl = "^1.0.3"
unidecode = "^1.3.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
